app/rag/retriever.py
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI
from langchain_openai import OpenAIEmbeddings
from langchain_community.vectorstores import FAISS

load_dotenv()

def get_rag_response(query: str):

    embeddings = OpenAIEmbeddings()

    vector_store = FAISS.load_local(
        "app/rag/faiss_index",
        embeddings,
        allow_dangerous_deserialization=True
    )

    retriever = vector_store.as_retriever(search_kwargs={"k": 3})

    docs = retriever.get_relevant_documents(query)

    context = "\n\n".join([doc.page_content for doc in docs])

    llm = ChatOpenAI(model="gpt-4o-mini", temperature=0.3)

    prompt = f"""
You are a cybersecurity expert.

Using the context below, answer the user's question clearly and professionally.

Context:
{context}

Question:
{query}

Answer:
"""

    response = llm.invoke(prompt)

    return response.content
